third_party/blink/web_tests/http/tests/htmlimports/cross-origin-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
CONSOLE ERROR: Access to imported resource at 'http://localhost:8080/htmlimports/resources/hello.html' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
PASS target.import is null
PASS successfullyParsed is true

TEST COMPLETE
